- Extracted inner class that creates parse tree and created new parser package - Started new specs for sqlbuilder package
