+++ src/main/java/chan/shundat/albert/sqlbuilder/parser/StringLiteralParseToken.java
@@ -1,0 +1,14 @@
+package chan.shundat.albert.sqlbuilder.parser;
+
+import chan.shundat.albert.sqlbuilder.SqlStringUtils;
+
+public class StringLiteralParseToken extends ParseToken {
+	private String value;
+
+	public String getValue() { return value; }
+	
+	public StringLiteralParseToken(String str) {
+		super(SqlStringUtils.createLiteralToken(str), false);
+		value = str;
+	}
+}
